requirements.txt
argon2-cffi==23.1.0
cryptography==42.0.4
dill==0.3.8
tqdm==4.66.2
